import books with the new structure

+++ database/migrations/2022_01_17_183434_create_comments_table.php
@@ -16,8 +16,8 @@ class CreateCommentsTable extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
             $table->text('body');
-            $table->integer('user_id')->unsigned();
-            $table->integer('book_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
+            $table->unsignedBigInteger('book_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('user_id')
